feat(infra): implement JPA repository and database config
